# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi System Laragpt",
        "description": "Descripción de mi API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Show All Companies",
                "operationId": "54f966c056050f9db4fff6ee5a45a76a",
                "responses": {
                    "200": {
                        "description": "List of Companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "My Name"
                                                    },
                                                    "image_path": {
                                                        "type": "string",
                                                        "example": "https://randomuser.me/api/portraits/women/23.jpg"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "New York, NY"
                                                    },
                                                    "industry": {
                                                        "type": "string",
                                                        "example": "Finance"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/companies?page=1"
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/companies?page=3"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/companies?page=2"
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "24"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Almacena un nuevo company en el sistema.",
                "operationId": "00e2db7b0f71c32b377f4dfc4dd685a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "industry",
                                    "user_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "My Name"
                                    },
                                    "image_path": {
                                        "type": "string",
                                        "example": "https://randomuser.me/api/portraits/women/23.jpg"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York, NY"
                                    },
                                    "industry": {
                                        "type": "string",
                                        "example": "Finance"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Entered Successfully"
                                        },
                                        "last_insert_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation exception"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Show Specific Company",
                "operationId": "01c261c039e4b27beeb888ef46bd86db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "My Name"
                                        },
                                        "image_path": {
                                            "type": "string",
                                            "example": "https://randomuser.me/api/portraits/women/23.jpg"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "New York, NY"
                                        },
                                        "industry": {
                                            "type": "string",
                                            "example": "Finance"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No exists company with id : #"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies/{company}": {
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update Exist Company in Database",
                "operationId": "17ee846405381b70d408b7532b44601a",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID del company a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "industry",
                                    "user_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "My Name"
                                    },
                                    "image_path": {
                                        "type": "string",
                                        "example": "https://randomuser.me/api/portraits/women/23.jpg"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York, NY"
                                    },
                                    "industry": {
                                        "type": "string",
                                        "example": "Finance"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Update Successfully"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company no encontrado"
                    },
                    "405": {
                        "description": "Validation exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation exception"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Remove Company From System",
                "operationId": "815006ac9bb0fb97b1cad53b1a620652",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID del company a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Record Delete Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company no encontrado"
                    }
                }
            }
        },
...
    }
}